% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendText.R
\name{sendText}
\alias{sendText}
\title{send a given text to a phone as SMS using Twilio API}
\usage{
sendText(strText, toNumber = NULL, fromNumber = NULL, sid = NULL,
  token = NULL)
}
\arguments{
\item{strText}{a string - the message to send}

\item{toNumber}{the target number of the SMS}

\item{fromNumber}{the Twilio assigned number for the user}

\item{sid}{Twilio account SID}

\item{token}{Twilio authorization token}
}
\value{
HTTP response from the POST method from Twilio API
}
\description{
send a string as text message using Twilio API service
}
\details{
If the parameters fromNumber, sid or token are skipped, global
variables are searched; which can be set using the Sys.setenv() function
call with parameters as twilio_number, twilio_sid and twilio_token
respectivly. Example Sys.setenv(twilio_number="yourtwilionumberasstring").
The function passes on the http message from the API post call and it is up
to the user to parse and interpret the response.
}

